Add UI and minor alterations to algorithm
A main application added which generates an algorithm then plays out a board using that algorithm.
Application uses scalafx as UI.
Board size reduced from 100x100 to 10x10.
Picking random members is now calculated based on picking from a random 10% of the population (rather than a flat count of 15).
The ratio to mutate is reduced from 20% to 10%.
Default starting coord for robot is now close to centre of board (rather than 0,0).

The UI code in main.scala is a bit of a mess, but it works ok. Next step, clean up/refactor UI code.
Also, iteration counts reduced to small numbers so things run faster. They need to be increased (to around 50-200) to see effective results.