Add and start using State from _Func Prog in Scala_
Add a version of State from ch. 6 of Chiusano, _Functional Programming in Scala_
Start using it in Population.evolveNewMember.
Currently there are multiple versions of the function, original, flatMap, and for-comprehension versions.
Slightly tweat some of the algorithm parameters, and the running speed in the window.

Currently algorithm is not very effective, but does move in right direction.